BRJ V4.5
-Add Healthbar
-Change main attack to  melee -> ranged
-Change Tematic to Pencil Case Battles

+++ NoShiftMe/Assets/Scripts/Bosses/Novice.cs
@@ -30,7 +30,7 @@ public class Novice : MonoBehaviour
     
     //Stats
     public int health;
-    public int contactDamage;
+    public float contactDamage;
     
     private void Start()
     {
